requirements.txt
asgiref==3.4.1
async-timeout==4.0.2
atomicwrites==1.4.1
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
coverage==6.4.2
cryptography==36.0.2
Deprecated==1.2.13
Django==4.0
django-cors-headers==3.11.0
django-crontab==0.7.1
django-debug-toolbar==3.3.0
django-hint==0.2.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-api-key==2.0.0
djangorestframework-simplejwt==4.8.0
drf-yasg==1.20.0
execnet==1.9.0
flake8==6.0.0
hiredis==2.0.0
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
mccabe==0.7.0
model-bakery==1.7.0
openapi-codec==1.3.2
packaging==21.0
Pillow==8.4.0
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.1.0
pyparsing==2.4.7
pyreadme==0.3
pytest==7.1.2
pytest-cov==3.0.0
pytest-django==4.5.2
pytest-forked==1.4.0
pytest-lazy-fixture==0.6.3
pytest-xdist==2.5.0
python-decouple==3.6
pytz==2021.3
redis==4.3.1
requests==2.27.1
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==3.10.0.2
tzdata==2022.1
uritemplate==3.0.1
urllib3==1.26.8
wrapt==1.14.1
